software patch one

+++ backup.sh
@@ -8,9 +8,17 @@ cp -f /etc/login.defs /CyberPatriot-Linux-Tools/old_files/
 cp -f /etc/sysctl.conf /CyberPatriot-Linux-Tools/old_files/
 cp -f /etc/rsyslog.d/50-default.conf /CyberPatriot-Linux-Tools/old_files/
 
-cp -f /etc/issue /CyberPatriot-Linux-Tools/old_files/
-cp -f /etc/issue.net /CyberPatriot-Linux-Tools/old_files/
-cp -f /etc/motd /CyberPatriot-Linux-Tools/old_files/
+if [ -f "/etc/issue" ]; then
+   cp -f /etc/issue /CyberPatriot-Linux-Tools/old_files/
+fi
+
+if [ -f "/etc/issue.net" ]; then
+   cp -f /etc/issue.net /CyberPatriot-Linux-Tools/old_files/
+fi
+
+if [ -f "/etc/motd" ]; then
+   cp -f /etc/motd /CyberPatriot-Linux-Tools/old_files/
+fi
 
 cp -f /etc/audit/audit.rules /CyberPatriot-Linux-Tools/old_files/
 cp -f /etc/audit/auditd /CyberPatriot-Linux-Tools/old_files/
@@ -18,7 +26,10 @@ cp -f /etc/audit/auditd /CyberPatriot-Linux-Tools/old_files/
 cp -f /etc/aide/aide.conf /CyberPatriot-Linux-Tools/old_files/
 
 cp -f /etc/profile /CyberPatriot-Linux-Tools/old_files/
-cp -f /etc/bashrc /CyberPatriot-Linux-Tools/old_files/
+
+if [ -f "/etc/bashrc" ]; then
+   cp -f /etc/bashrc /CyberPatriot-Linux-Tools/old_files/
+fi
 
 cp -f /etc/hosts.allow /CyberPatriot-Linux-Tools/old_files/
 cp -f /etc/hosts.deny /CyberPatriot-Linux-Tools/old_files/
